+++ src/App.js
@@ -2,24 +2,22 @@ import React, { Suspense } from "react";
 import { BrowserRouter, Switch, Route, Redirect } from "react-router-dom";
 import { PrivateRoute } from "./Components/Helpers/PrivateRoute";
 import Splashpage from "./Components/Splashpage";
-import Signup from "./Components/Signup";
 import ResetPassword from "./Components/Login/ResetPassword";
 import ForgotPassword from "./Components/Login/ForgotPassword";
-import OrgSelect from "./Components/OrgSelect";
+import OrgSelect from "./Components/Organizations/OrgSelect";
 import Dashboard from "./Components/Dashboard";
 import BoilerplatesIndex from "./Components/Boilerplates/BoilerplatesIndex";
 import CategoriesIndex from "./Components/Categories/CategoriesIndex";
-import Organizations from "./Components/Organizations";
+import Organizations from "./Components/Organizations/Organizations";
 import GrantsIndex from "./Components/Grants/GrantsIndex";
 import FundingOrgsIndex from "./Components/FundingOrgs/FundingOrgsIndex";
 import BoilerplatesShow from "./Components/Boilerplates/BoilerplatesShow";
-import OrganizationsShow from "./Components/OrganizationsShow";
-import ReportsShow from "./Components/ReportsShow";
+import OrganizationsShow from "./Components/Organizations/OrganizationsShow";
+import ReportsShow from "./Components/Reports/ReportsShow";
 import GrantsShow from "./Components/Grants/GrantsShow";
-import ReportsFinalizeShow from "./Components/ReportsFinalizeShow";
 import FundingOrgNew from "./Components/FundingOrgs/FundingOrgNew";
 import BoilerplatesNew from "./Components/Boilerplates/BoilerplatesNew";
-import OrganizationsNew from "./Components/OrganizationsNew";
+import OrganizationsNew from "./Components/Organizations/OrganizationsNew";
 import GrantsNew from "./Components/Grants/GrantsNew";
 import SectionsNew from "./Components/Sections/SectionsNew";
 import ReportsNew from "./Components/Reports/ReportsNew";
@@ -40,7 +38,6 @@ export default function App() {
             <Redirect to="/splashpage" />
           </Route>
           <Route path="/splashpage" component={Splashpage} />
-          <Route path="/signup" component={Signup} />
           <Route path="/reset_password" component={ResetPassword} />
           <Route path="/forgot_password" component={ForgotPassword} />
           <CurrentOrganizationProvider>
@@ -85,12 +82,6 @@ export default function App() {
                         "/organizations/:organizationId/grants/:grant_id/reports/:report_id"
                       }
                       component={ReportsShow}
-                    />
-                    <PrivateRoute
-                      path={
-                        "/organizations/:organizationId/grants/:grant_id/reports-finalize/:report_id"
-                      }
-                      component={ReportsFinalizeShow}
                     />
                     <PrivateRoute
                       path="/organizations/:organizationId/grants/:grant_id/reports-new"
